+++ agixt/provider/__init__.py
@@ -4,6 +4,58 @@ import pkg_resources
 import glob
 import os
 import inspect
+from DBConnection import (
+    session,
+    Provider,
+    ProviderSetting,
+)
+
+
+def import_providers():
+    providers = get_providers()
+    existing_providers = session.query(Provider).all()
+    existing_provider_names = [provider.name for provider in existing_providers]
+
+    for provider in existing_providers:
+        if provider.name not in providers:
+            session.delete(provider)
+
+    for provider_name in providers:
+        provider_options = get_provider_options(provider_name)
+
+        provider = session.query(Provider).filter_by(name=provider_name).one_or_none()
+
+        if provider:
+            print(f"Updating provider: {provider_name}")
+        else:
+            provider = Provider(name=provider_name)
+            session.add(provider)
+            existing_provider_names.append(provider_name)
+            print(f"Adding provider: {provider_name}")
+
+        for option_name, option_value in provider_options.items():
+            provider_setting = (
+                session.query(ProviderSetting)
+                .filter_by(provider_id=provider.id, name=option_name)
+                .one_or_none()
+            )
+            if provider_setting:
+                provider_setting.value = option_value
+                print(
+                    f"Updating provider setting: {option_name} for provider: {provider_name}"
+                )
+            else:
+                provider_setting = ProviderSetting(
+                    provider_id=provider.id,
+                    name=option_name,
+                    value=option_value,
+                )
+                session.add(provider_setting)
+                print(
+                    f"Adding provider setting: {option_name} for provider: {provider_name}"
+                )
+
+    session.commit()
 
 
 def get_providers():
